Separate AI deliverer.
